refactoring + fix data size calc in decoder
